Deny connection if user does not match annotation
If USER_BEARER_TOKEN is enabled, we check the X-Forwarded-User header
for a username and compare that to an annotation on the pod. If they
match, continue; otherwise do not allow access.

This is to support restricting access to only the creator when creating
a cloud-shell via the che workspace operator

Signed-off-by: Angel Misevski <amisevsk@redhat.com>